Move `GetOrNoneManager`
