+++ corehq/apps/case_search/models.py
@@ -18,6 +18,7 @@ from dimagi.utils.parsing import string_to_boolean
 from corehq.apps.case_search.exceptions import CaseSearchUserError
 from corehq.apps.case_search.filter_dsl import CaseFilterError
 from corehq.util.metrics.const import MODULE_NAME_TAG
+from corehq.util.models import GetOrNoneManager
 from corehq.util.quickcache import quickcache
 
 CLAIM_CASE_TYPE = 'commcare-case-claim'
@@ -214,18 +215,6 @@ def extract_search_request_config(request_dict):
     }
     criteria = criteria_dict_to_criteria_list(params)
     return CaseSearchRequestConfig(criteria=criteria, **kwargs_from_params)
-
-
-class GetOrNoneManager(models.Manager):
-    """
-    Adds get_or_none method to objects
-    """
-
-    def get_or_none(self, **kwargs):
-        try:
-            return self.get(**kwargs)
-        except self.model.DoesNotExist:
-            return None
 
 
 class FuzzyProperties(models.Model):
